+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [#354](https://github.com/loop-payments/prisma-lint/issues/354) Add `allowlist` to `model-name-grammatical-number`.
+
 ## 0.1.3 (2024-03-23)
 
 - Upgrade `@mrleebo/prisma-ast` to get new version that supports mapped enums.